datalake: move direct_table_creator to tests
It is not used in the main codebase, so move it to tests to avoid
confusion.
