+++ src/v/datalake/coordinator/tests/BUILD
@@ -11,7 +11,6 @@ redpanda_test_cc_library(
     include_prefix = "datalake/coordinator/tests",
     deps = [
         "//src/v/container:fragmented_vector",
-        "//src/v/datalake:table_creator",
         "//src/v/datalake/coordinator:file_committer",
         "//src/v/datalake/coordinator:state",
         "//src/v/datalake/coordinator:translated_offset_range",
@@ -31,7 +30,9 @@ redpanda_cc_gtest(
         ":state_test_utils",
         "//src/v/cluster",
         "//src/v/config",
+        "//src/v/datalake:catalog_schema_manager",
         "//src/v/datalake:logger",
+        "//src/v/datalake:record_schema_resolver",
         "//src/v/datalake/coordinator",
         "//src/v/datalake/coordinator:file_committer",
         "//src/v/datalake/coordinator:state",
@@ -60,6 +61,7 @@ redpanda_cc_gtest(
         "//src/v/cloud_io:remote",
         "//src/v/cloud_io/tests:scoped_remote",
         "//src/v/cloud_storage/tests:s3_imposter",
+        "//src/v/datalake:catalog_schema_manager",
         "//src/v/datalake:logger",
         "//src/v/datalake:table_definition",
         "//src/v/datalake/coordinator:iceberg_file_committer",
@@ -104,6 +106,8 @@ redpanda_cc_gtest(
     deps = [
         ":state_test_utils",
         "//src/v/cluster",
+        "//src/v/datalake:catalog_schema_manager",
+        "//src/v/datalake:record_schema_resolver",
         "//src/v/datalake/coordinator",
         "//src/v/datalake/coordinator:stm",
         "//src/v/raft/tests:stm_test_fixture",
