+++ src/v/datalake/coordinator/tests/coordinator_test.cc
@@ -10,12 +10,14 @@
 #include "cluster/data_migrated_resources.h"
 #include "cluster/topic_table.h"
 #include "config/mock_property.h"
+#include "datalake/catalog_schema_manager.h"
 #include "datalake/coordinator/coordinator.h"
 #include "datalake/coordinator/file_committer.h"
 #include "datalake/coordinator/state_machine.h"
 #include "datalake/coordinator/state_update.h"
 #include "datalake/coordinator/tests/state_test_utils.h"
 #include "datalake/logger.h"
+#include "datalake/record_schema_resolver.h"
 #include "raft/tests/raft_fixture.h"
 #include "random/generators.h"
 #include "test_utils/async.h"
