+++ src/v/datalake/coordinator/tests/state_machine_test.cc
@@ -10,9 +10,11 @@
 
 #include "cluster/data_migrated_resources.h"
 #include "cluster/topic_table.h"
+#include "datalake/catalog_schema_manager.h"
 #include "datalake/coordinator/coordinator.h"
 #include "datalake/coordinator/state_machine.h"
 #include "datalake/coordinator/tests/state_test_utils.h"
+#include "datalake/record_schema_resolver.h"
 #include "raft/tests/stm_test_fixture.h"
 
 using coordinator = std::unique_ptr<datalake::coordinator::coordinator>;
