+++ src/v/datalake/table_creator.cc
@@ -9,9 +9,6 @@
  */
 #include "datalake/table_creator.h"
 
-#include "datalake/record_translator.h"
-#include "datalake/table_id_provider.h"
-
 namespace datalake {
 
 std::ostream& operator<<(std::ostream& o, const table_creator::errc& e) {
@@ -23,45 +20,6 @@ std::ostream& operator<<(std::ostream& o, const table_creator::errc& e) {
     case table_creator::errc::shutting_down:
         return o << "table_creator::errc::shutting_down";
     }
-}
-
-direct_table_creator::direct_table_creator(
-  type_resolver& tr, schema_manager& sm)
-  : type_resolver_(tr)
-  , schema_mgr_(sm) {}
-
-ss::future<checked<std::nullopt_t, table_creator::errc>>
-direct_table_creator::ensure_table(
-  const model::topic& topic,
-  model::revision_id,
-  record_schema_components comps) const {
-    auto table_id = table_id_provider::table_id(topic);
-
-    std::optional<resolved_type> val_type;
-    if (comps.val_identifier) {
-        auto type_res = co_await type_resolver_.resolve_identifier(
-          comps.val_identifier.value());
-        if (type_res.has_error()) {
-            co_return errc::failed;
-        }
-        val_type = std::move(type_res.value());
-    }
-
-    auto record_type = default_translator{}.build_type(std::move(val_type));
-    auto ensure_res = co_await schema_mgr_.ensure_table_schema(
-      table_id, record_type.type);
-    if (ensure_res.has_error()) {
-        switch (ensure_res.error()) {
-        case schema_manager::errc::not_supported:
-            co_return errc::incompatible_schema;
-        case schema_manager::errc::failed:
-            co_return errc::failed;
-        case schema_manager::errc::shutting_down:
-            co_return errc::shutting_down;
-        }
-    }
-
-    co_return std::nullopt;
 }
 
 } // namespace datalake
